requirements.txt
Flask == 3.0.3
metapub == 0.5.12

alembic
APScheduler
bcrypt
certifi
charset-normalizer
click
colorama
cryptography
Flask-APScheduler
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
flask-marshmallow
Flask-Migrate
flask-sqlalchemy
greenlet
gunicorn
idna
importlib-metadata
importlib-resources
isort
itsdangerous
jinja2
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy
packaging
pip
pyclean
PyJWT
PyMySQL
python-dateutil
python-dotenv
python-json-logger
pytz
requests
setuptools
six
SQLAlchemy
typing-extensions
tzdata
tzlocal
urllib3
waitress
Werkzeug
zipp